requirements.txt
# Core ML & NLP
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
onnxruntime>=1.15.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API & Web
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic[dotenv]>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0

# Web Scraping & Parsing
beautifulsoup4>=4.12.0
requests>=2.31.0

# Firebase Integration
firebase-admin>=6.2.0

# Visualization & Logging
matplotlib>=3.7.0
rich>=13.0.0
tqdm>=4.65.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
